JENKINS-28640: Quotation marks in commit msg:
Only related to the accumulated strategy.

Temporary fix for the bug with quotation marks in the commit message
leading to merge failure because of not properly escaped and formatted
command line arguments.

Fix is temporary replacing " (quotation marks) with ' (plings).

The squashed strategy is not affected, as the git command automatically
generated the commit message, where as the plugin creates a look-a-like
squash message, slightly modified, is using accumulated strategy.

Other improvements:

- fixed hard-coded dependency on the maven-hpi-plugin in the pom file,
  new version works now
- cleaned few tests from wrong comments, and not clearly formatted error
  messages if the tests failed
- renamed test suite for commit message prolems related to quotation
  marks to special chars test suite
- shared a useful method from one test suite through test utils class
  that can pretty print console output from the job to use for
verification in the tests.