JENKINS-29369 Fix locale dependent accumulated msg
The accumulated commit message is now using english locale when
formatting date and time. The commit message is generated by the plugin
and formatted to look like the squash commit message except for the
header line.

Tests have been improved to include detailed commit message asserts and
extend to include few more tests for squashed also.
Each test now verifies that the formatting of dates are correct for the
tests regarding special charecters and custom integration branch (those
were the newer tests we added so the follow a better design).

Other improvements and refactoring:

- removed un-used code in squash commit strategy
- removed debug print of squash commit strategy to build console
- add debug printing of exception during integration to the build
  console
- fixed minor bug in test utils regarding printing console which is null
- test base class for static git repos extended to print out cleaning
  repos and temp files